experiments/mbu_netpp/requirements.txt
torch
torchvision
segmentation-models-pytorch
timm
albumentations
opencv-python
numpy
PyYAML
scikit-image
Pillow
openpyxl
